Create RxSchedulerExtension to use it in JUnit5 tests
