+++ app-example/src/test/java/com/badoo/ribs/example/RxSchedulerRule.kt
@@ -1,20 +1,12 @@
 package com.badoo.ribs.example
 
-import io.reactivex.Scheduler
-import io.reactivex.android.plugins.RxAndroidPlugins
-import io.reactivex.functions.Function
-import io.reactivex.plugins.RxJavaPlugins
-import io.reactivex.schedulers.Schedulers
 import org.junit.rules.TestRule
 import org.junit.runner.Description
 import org.junit.runners.model.Statement
-import java.util.concurrent.Callable
 
 class RxSchedulerRule : TestRule {
-    private val scheduler = Schedulers.trampoline()
-    private val schedulerFunction: Function<Scheduler, Scheduler> = Function { scheduler }
-    private val schedulerFunctionLazy: Function<Callable<Scheduler>, Scheduler> =
-        Function { scheduler }
+
+    private val extension = RxSchedulerExtension()
 
     override fun apply(
         base: Statement,
@@ -23,15 +15,9 @@ class RxSchedulerRule : TestRule {
         return object : Statement() {
             @Throws(Throwable::class)
             override fun evaluate() {
-                RxAndroidPlugins.reset()
-                RxAndroidPlugins.setInitMainThreadSchedulerHandler(schedulerFunctionLazy)
-                RxJavaPlugins.reset()
-                RxJavaPlugins.setIoSchedulerHandler(schedulerFunction)
-                RxJavaPlugins.setNewThreadSchedulerHandler(schedulerFunction)
-                RxJavaPlugins.setComputationSchedulerHandler(schedulerFunction)
+                extension.beforeEach(null)
                 base.evaluate()
-                RxAndroidPlugins.reset()
-                RxJavaPlugins.reset()
+                extension.afterEach(null)
             }
         }
     }
